+++ androidApp/src/androidMain/kotlin/at/asitplus/wallet/app/android/NdefDeviceEngagementService.kt
@@ -285,7 +285,7 @@ class NdefDeviceEngagementService : HostApduService() {
         // If the reader hasn't connected by the time NFC interaction ends, make sure we only
         // wait for a limited amount of time.
         if (presentationStateModel.state.value == PresentationStateModel.State.CONNECTING) {
-            val timeout = 15.seconds
+            val timeout = settings.connectionTimeout
             Napier.i("NdefDeviceEngagementService: Reader hasn't connected at NFC deactivation time, scheduling $timeout timeout for closing")
             disableEngagementJob = CoroutineScope(Dispatchers.IO).launch {
                 delay(timeout)