+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/presentation/TransferSettings.kt
@@ -1,5 +1,8 @@
 package at.asitplus.wallet.app.common.presentation
 
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.seconds
+
 data class TransferSettings(
     val presentmentNegotiatedHandoverPreferredOrder: List<String> = listOf(
         "ble:central_client_mode:",
@@ -12,4 +15,5 @@ data class TransferSettings(
     val presentmentNfcDataTransferEnabled: Boolean = true,
     val readerBleL2CapEnabled: Boolean = true,
     val presentmentAllowMultipleRequests: Boolean = false,
+    val connectionTimeout: Duration = 15.seconds
 )