+++ CHANGELOG.md
@@ -1,4 +1,4 @@
- * Add NFC device engagement/retrieval and Bluetooth device retrieval mechanisms
+ * Android: Add NFC device engagement/retrieval and Bluetooth device retrieval mechanisms
 
 # Release 5.4.16
  * Recognize DCQL queries for consent preview and show dummy consent screen